requirements.txt
fastapi
uvicorn
python-dotenv
web3
sqlalchemy
psycopg2-binary
pandas
pymysql